Spawned objects inherit speed now (not just T)
This is not yet what #41 was about, but at least a nod to that.
Also: introducing arbitrary momentum (speed) to the world is kinda
crucial, as the raw "real" physics has been such a boring drag...:
everything has always just collapsed into a black hole eventually,
and that was it. :)

+ Exhaust bursts are configurable now, and the sprinkles are also more
  agile ("reactive") by default: nice whirling formations can be created
  easily out of them, and the plume of an acceleration burn now visibly
  reacts with itself, too.
+ Updated Help panel