Revamped audio backend (-> sound players/channels)
A fixed number of max. parallel sounds are supported, which go round-robin
to reuse the oldest one on congestion. (But... -> Need to prioritize and
"protect" looped sounds later!)

Also had to wrap the sound buffers in shared_ptrs to be able to add them
to containers, otherwise e.g. verctor grow kept destructing & reconstructing
them incessantly, which was "not good" while also being played. :)

+ Close #24: A "plip" sound for spawning (three diff. ones)
+ Close #417: Glassy jingling sound to chemtrail release