Close #458: Move render/draw to a polymorphic View
The View needs to be polymorphized, integrated into the app (hierarchy),
and the app has to play along with construction, class (member/base)
layouts, adjusted calls (to it, instead of the renderer) etc.

+ Also some #461: Most of the renderer is now merged into the OON View.
+ Some previously SFML-specific app-level stuff (via their dep. on the
  SFML-specific Renderer) have been moved from OON_sfml to OON, finally.

FOLLOW-UP NOTE regarding the mysterious slowdown in the prev. change:

It's gone. :-o Now it's back to >160 FPS windowed (~200 full-screen).
(The CPU throttling is at the same "slowest" setting as yesterday.)