Close #528: Templatize the model number type
The fundamental type for all the physics params used to be hardcoded as
float. Now it can be changed (with a #define), and it's double by default.

Surprisingly, the overall FPS hasn't changed *at all* after this change,
even on my oldish, slow-CPU TOSHIBA laptop! :-o

(NOTE: snapshot save/load can only handle the old format with floats yet!
See #561 for that!)

+ Minor random corollary changes here and there.